+++ components/telemetry-operator/internal/utils/fs.go
