+++ components/telemetry-operator/internal/webhook/logpipeline/webhook_suite_test.go
@@ -25,10 +25,10 @@ import (
 	"testing"
 	"time"
 
+	utilsmocks "github.com/kyma-project/kyma/components/telemetry-operator/internal/utils/mocks"
+
 	"github.com/kyma-project/kyma/components/telemetry-operator/internal/webhook/logpipeline/mocks"
 	validationmocks "github.com/kyma-project/kyma/components/telemetry-operator/internal/webhook/logpipeline/validation/mocks"
-	utilsmocks "github.com/kyma-project/kyma/components/telemetry-operator/internal/webhook/utils/mocks"
-
 	"github.com/prometheus/client_golang/prometheus"
 	"sigs.k8s.io/controller-runtime/pkg/metrics"
